=last changes

+++ src/invader.cpp
@@ -12,46 +12,77 @@
 #include "Sound.cpp"
 #include "SoundPlayer.cpp"
 
-#include "invader_images.cpp"
-#include "invader_sounds.cpp"
-#include "Stage1.cpp"
+#include "Invader_images.cpp"
+#include "Invader_sounds.cpp"
+#include "Invader_entities.cpp"
+#include "Invader_stage1.cpp"
+
+Tilemap *Invader::background=NULL;
+EntityManager *Invader::entity_manager=NULL;
+bool Invader::initialized=false;
+
+void Invader::init(){
+    if(Invader::initialized == true){
+        fprintf(stderr, "Invader::init : Invader already initialized!");
+        return;
+    }
+
+    Invader::background = new Tilemap();
+    Invader::entity_manager = new EntityManager();
+
+    Graphics::createWindow(INVADER_WINDOW_WIDTH,INVADER_WINDOW_HEIGHT);
+    SoundPlayer::init();
+    loadImages();
+    loadSounds();
+
+    Invader::initialized = true;
+    
+}
+
+void Invader::deInit(){
+    if(Invader::initialized == false){
+        fprintf(stderr, "Invader::deInit : Invader not initialized!");
+        return;
+    }
 
 
-void deInit(){
-	
+    delete Invader::background;
+    delete Invader::entity_manager;
+
 	// destroy managers...
 	Font::destroy();
 	Image::destroy();
 	Sound::destroy();
-    Tilemap::destroy();
-    EntityManager::destroy();
-
-
 	Graphics::deInit();
 	SoundPlayer::deInit();
 
 	// unload SDL
 	SDL_Quit();
+
+    Invader::initialized = false;
 }
 
 int main(int argc, char *argv[]){
-    Graphics::createWindow(INVADER_WINDOW_WIDTH,INVADER_WINDOW_HEIGHT);
+
+
+    Invader::init();
+
     Stage1 stage1;
-
-    SoundPlayer::init();
-
-    loadImages();
-    loadSounds();
-    
     stage1.load();
     
+
+    int idx_current_image=0;
+    Image *current_image=Image::get(image_files[idx_current_image].file_name);
+
+    int idx_current_sound=0;
+    Sound *current_sound=Sound::get(sound_files[idx_current_sound]);
 
     do{
 
         // clear screen...
 		Graphics::clear(0,0,0);
 
-        /*if(current_image!=NULL){
+        if(current_image!=NULL){
             Graphics::drawImage(current_image,current_image->getWidth()>>1,current_image->getHeight()>>1);
         }
 
@@ -69,7 +100,7 @@ int main(int argc, char *argv[]){
             }
             idx_current_sound=(idx_current_sound+1)%sound_files.size();            
             current_sound=Sound::get(sound_files[idx_current_sound]);
-        }*/
+        }
